+++ src/test/java/ru/levelup/homework_number_five/task_number_one/AbstractBaseCalculator.java
@@ -1,19 +1,19 @@
-package ru.levelup.homework_number_three.task_number_one;
+package ru.levelup.homework_number_five.task_number_one;
 
-import org.testng.annotations.AfterTest;
-import org.testng.annotations.BeforeTest;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
 import ru.levelup.qa.at.calculator.Calculator;
 
 public abstract class AbstractBaseCalculator {
 
   protected Calculator calculator;
 
-  @BeforeTest
+  @BeforeMethod
   public void beforeTestClass() {
     calculator = new Calculator();
   }
 
-  @AfterTest
+  @AfterMethod
   public void beforeTestMethod() {
     calculator = null;
   }
