+++ src/test/java/ru/levelup/homework_number_five/task_number_one/CalculatorSumTest.java
@@ -1,4 +1,4 @@
-package ru.levelup.homework_number_three.task_number_one;
+package ru.levelup.homework_number_five.task_number_one;
 
 import jdk.jfr.Description;
 import org.testng.Assert;
@@ -9,7 +9,7 @@ public class CalculatorSumTest extends AbstractBaseCalculator{
   @Test
   @Description("Тест на сложение")
   public void sumTest(){
-    long actual = calculator.sub(35,9);
+    long actual = calculator.sum(35,9);
     long expected = 35 + 9;
     Assert.assertEquals(actual, expected, "Incorrect result");
   }
