+++ src/test/resources/ru.levelup.homework_number_five.task_number_one/CalculatorOthersTestSuite.xml
@@ -1,0 +1,12 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+
+<suite name="Calculator Multi and Divide Test Suite">
+  <test name="CalculatorAllTest">
+    <classes>
+      <class name="ru.levelup.homework_number_five.task_number_one.CalculatorPowTest"/>
+      <class name="ru.levelup.homework_number_five.task_number_one.CalculatorSqrtTest"/>
+      <class name="ru.levelup.homework_number_five.task_number_one.CalculatorIsPositiveTest"/>
+      <class name="ru.levelup.homework_number_five.task_number_one.CalculatorIsNegativeTest"/>
+    </classes>
+  </test>
+</suite>
