+++ src/test/resources/ru.levelup.homework_number_five.task_number_one/CalculatorSumSubTestSuite.xml
@@ -1,0 +1,10 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+
+<suite name="Calculator Multi and Divide Test Suite">
+  <test name="CalculatorAllTest">
+    <classes>
+      <class name="ru.levelup.homework_number_five.task_number_one.CalculatorSumTest"/>
+      <class name="ru.levelup.homework_number_five.task_number_one.CalculatorSubTest"/>
+    </classes>
+  </test>
+</suite>
